replaced config with cache, moved argparser into separate module
