+++ .gitignore
@@ -1,6 +1,9 @@
 __pycache__
 pyvenv.cfg
-venv
 .vscode
 .cache
 .pytest_cache
+
+# virtual environments for win and linux
+wenv
+lenv
